ATR-762: moved PX1093 unit tests to the new location and adjusted them to the changes in analysis
