+++ src/Acuminator/Acuminator.Tests/Tests/StaticAnalysis/DeclarationAnalysisGraphAndDac/Sources/GraphDeclarationTypeParameter/Graph_Bad.cs
